requirements.txt
asgiref==3.9.1
Django>=5.2,<6.0 
django-background-tasks==1.2.8
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
psycopg[binary]==3.2.3     
psycopg>=3.1
Pillow>=9.0.0
daphne>=4.0
channels>=4.0
whitenoise
channels==4.3.0
stripe==13.2.0
resend==2.19.0
channels-redis>=4.2.0
redis>=5.0.0
google-generativeai>=0.8.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy==1.26.4
